Added some comparison operators, and updated test.add
First, I'm not entirely sure if I represented n1 and n2 as a int, spring or float correctly. Second, I understand the IsNot expression using "/=" to compare the n1 and n2. Which I also used for NotEqual since it said to in the issues, however, why wouldn't it just use "!=" ?